Adopt a more declarative approach to testing
