+++ .gitignore
@@ -3,5 +3,5 @@ Manifest-*.toml
 .swp
 .DS_Store
 docs/build
-test/reference_tests/*_rec.png
-test/reference_tests/*_diff.png
+test/plots/images/*_rec.png
+test/plots/images/*_diff.png
