+++ test/plots/images/autocorplot_ref.png
