+++ test/plots/images/makie_plot_ref.png
