+++ test/plots/images/meanplot_ref.png
