+++ test/plots/images/mixeddensity_float_ref.png
