+++ test/plots/images/mixeddensity_int_ref.png
