+++ test/plots/images/mmixeddensity_float_ref.png
