+++ test/plots/images/mmixeddensity_int_ref.png
