+++ test/plots/images/mrankplot_overlay_ref.png
