+++ test/plots/images/mtraceplot_ref.png
