+++ test/plots/images/pairplot_pooled_ref.png
