+++ test/plots/images/pairplot_ref.png
