+++ test/plots/images/plots_plot_ref.png
