+++ test/plots/images/rankplot_overlay_ref.png
