+++ test/plots/images/rankplot_ref.png
