+++ test/plots/images/traceplot_ref.png
